Restructure otel integration
